--- packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/private/format/util.ts
@@ -1,86 +1,0 @@
-import type * as cxapi from '@aws-cdk/cx-api';
-import * as cxschema from '@aws-cdk/cloud-assembly-schema';
-import { Writable } from 'stream';
-import { TemplateDiff } from '@aws-cdk/cloudformation-diff';
-
-/*
- * Custom writable stream that collects text into a string buffer.
- * Used on classes that take in and directly write to a stream, but
- * we intend to capture the output rather than print.
- */
-export class StringWriteStream extends Writable {
-  private buffer: string[] = [];
-
-  constructor() {
-    super();
-  }
-
-  _write(chunk: any, _encoding: string, callback: (error?: Error | null) => void): void {
-    this.buffer.push(chunk.toString());
-    callback();
-  }
-
-  toString(): string {
-    return this.buffer.join('');
-  }
-}
-
-export function buildLogicalToPathMap(stack: cxapi.CloudFormationStackArtifact) {
-  const map: { [id: string]: string } = {};
-  for (const md of stack.findMetadataByType(cxschema.ArtifactMetadataEntryType.LOGICAL_ID)) {
-    map[md.data as string] = md.path;
-  }
-  return map;
-}
-
-
-export function logicalIdMapFromTemplate(template: any) {
-  const ret: Record<string, string> = {};
-
-  for (const [logicalId, resource] of Object.entries(template.Resources ?? {})) {
-    const path = (resource as any)?.Metadata?.['aws:cdk:path'];
-    if (path) {
-      ret[logicalId] = path;
-    }
-  }
-  return ret;
-}
-
-/**
- * Remove any template elements that we don't want to show users.
- * This is currently:
- * - AWS::CDK::Metadata resource
- * - CheckBootstrapVersion Rule
- */
-export function obscureDiff(diff: TemplateDiff) {
-  if (diff.unknown) {
-    // see https://github.com/aws/aws-cdk/issues/17942
-    diff.unknown = diff.unknown.filter(change => {
-      if (!change) {
-        return true;
-      }
-      if (change.newValue?.CheckBootstrapVersion) {
-        return false;
-      }
-      if (change.oldValue?.CheckBootstrapVersion) {
-        return false;
-      }
-      return true;
-    });
-  }
-
-  if (diff.resources) {
-    diff.resources = diff.resources.filter(change => {
-      if (!change) {
-        return true;
-      }
-      if (change.newResourceType === 'AWS::CDK::Metadata') {
-        return false;
-      }
-      if (change.oldResourceType === 'AWS::CDK::Metadata') {
-        return false;
-      }
-      return true;
-    });
-  }
-}
